+++ src/main/kotlin/com/github/diekautz/ideplugin/extensions/EditorExtensions.kt
@@ -5,10 +5,13 @@ import com.intellij.codeInsight.highlighting.HighlightManager
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.editor.LogicalPosition
+import com.intellij.openapi.fileEditor.FileEditor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiElement
+import com.intellij.util.ui.ImageUtil
 import java.awt.Point
+import java.awt.image.BufferedImage
 import javax.swing.SwingUtilities
 
 fun Editor.xyScreenToLogical(point: Point): LogicalPosition {
@@ -42,6 +45,12 @@ fun Editor.highlightElementGazePoints(seen: Map<PsiElement, Double>, project: Pr
             project
         )
     }
+}
+
+fun FileEditor.screenshot(): BufferedImage {
+    val bufferedImage = ImageUtil.createImage(component.height, component.width, BufferedImage.TYPE_INT_RGB)
+    component.paintAll(bufferedImage.graphics)
+    return bufferedImage
 }
 
 fun EditorFactory.removeAllHighlighters() {