+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 
+- Added saving of an image of all editors when saving recording data
 - Interrupt user and prompt for answer/code while recording
 - Added configuration option for interrupt
 - Added action for highlighting gaze elements